+++ modules/router/src/test/collateral.spec.ts
@@ -5,23 +5,31 @@ import {
   getRandomBytes32,
   mkAddress,
   RestServerNodeService,
-  getBalanceForAssetId,
+  mkPublicIdentifier,
+  getTestLoggers,
 } from "@connext/vector-utils";
 import Sinon from "sinon";
-import pino from "pino";
 import { BigNumber } from "@ethersproject/bignumber";
 import { AddressZero } from "@ethersproject/constants";
-import { INodeService, Result } from "@connext/vector-types";
+import { INodeService, Result, UpdateType } from "@connext/vector-types";
+import { parseEther } from "@ethersproject/units";
 
 import { config } from "../config";
-import { requestCollateral } from "../collateral";
-import { getRebalanceProfile } from "../services/config";
+import { adjustCollateral, justInTimeCollateral, requestCollateral } from "../services/collateral";
+import * as configService from "../services/config";
+import { CollateralError } from "../errors";
 
-const logger = pino({ level: config.logLevel });
+const testName = "Collateral";
+const { log } = getTestLoggers(testName, config.logLevel as any);
+const chainId = parseInt(Object.keys(config.chainProviders)[0]);
 
-describe("Collateral", () => {
+describe(testName, () => {
   let node: Sinon.SinonStubbedInstance<RestServerNodeService>;
   let chainReader: Sinon.SinonStubbedInstance<VectorChainReader>;
+  let getRebalanceProfile: Sinon.SinonStub;
+
+  const ethProfile = config.rebalanceProfiles.find((p) => p.chainId === chainId && p.assetId === AddressZero);
+  const routerPublicIdentifier = mkPublicIdentifier("vectorRRR");
 
   beforeEach(async () => {
     node = Sinon.createStubInstance(RestServerNodeService);
@@ -37,171 +45,385 @@ describe("Collateral", () => {
         [1337]: { waitForTransaction: () => Promise.resolve({ logs: [] }) } as any,
       }),
     );
+
+    getRebalanceProfile = Sinon.stub(configService, "getRebalanceProfile");
+    getRebalanceProfile.returns(Result.ok(ethProfile));
   });
 
   afterEach(() => {
-    // TODO: why doesnt this work
     Sinon.restore();
     Sinon.reset();
   });
 
-  // it("should get profiles for different assetIds", async () => {
-  //   const channel = createTestChannelState("create").channel;
-  //   const rebalanceProfileEthRes = await getRebalanceProfile(channel.networkContext.chainId, AddressZero);
-  //   const profileEth = rebalanceProfileEthRes.getValue();
-  //   expect(profileEth.assetId).eq(AddressZero);
+  describe("justInTimeCollateral", () => {
+    const transferAmount = parseEther("0.001");
 
-  //   const tokenProfile = config.rebalanceProfiles.find(
-  //     (prof) => prof.chainId === channel.networkContext.chainId && prof.assetId !== AddressZero,
-  //   );
-  //   const rebalanceProfileTokenRes = await getRebalanceProfile(channel.networkContext.chainId, tokenProfile.assetId);
-  //   const profileToken = rebalanceProfileTokenRes.getValue();
-  //   expect(profileToken.assetId).eq(tokenProfile.assetId);
-  // });
-  it("should fail if getRebalanceProfile fails", async () => {});
-  it("should fail if requestedAmount > reclaimThreshold", async () => {});
-  it("should fail if it cannot get the chainProviders", async () => {});
-  it("should fail if it cannot get a providerUrl", async () => {});
-  it("should fail if it cannot get the onchain balance", async () => {});
+    it("should do nothing if there is sufficient balance for payment", async () => {
+      const { channel } = createTestChannelState(UpdateType.deposit, {
+        alice: mkAddress("0xaaa"),
+        aliceIdentifier: routerPublicIdentifier,
+        assetIds: [AddressZero],
+        balances: [{ to: [mkAddress("0xaaa"), mkAddress("0xbbb")], amount: [transferAmount.mul(3).toString(), "0"] }],
+      });
+      const res = await justInTimeCollateral(
+        channel,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader,
+        log,
+        transferAmount.toString(),
+      );
+      expect(res.getError()).to.be.undefined;
+      expect(res.getValue()).to.be.undefined;
+      expect(node.sendDepositTx.callCount).to.be.eq(0);
+    });
 
-  describe("should work", () => {
-    it("if profile.target > requestedAmount (deposits larger of two)", async () => {});
-    it("if requestedAmount > profile.target (deposits larger of two)", async () => {});
-    it("if transferAmount provided", async () => {});
-    it("if no collateral needed", async () => {});
-    it("if there is only offchain reconciliation needed (no deposit sent onchain)", async () => {});
-    it("if an onchain deposit is needed", async () => {});
-    it("if requested amount higher than the target", async () => {});
-  });
+    it("should fail if it cannot get the collateral profile", async () => {
+      getRebalanceProfile.returns(Result.fail(new Error("fail")));
+      const { channel } = createTestChannelState(UpdateType.deposit);
+      const res = await justInTimeCollateral(
+        channel,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader,
+        log,
+        transferAmount.toString(),
+      );
+      expect(res.getError().message).to.be.eq(CollateralError.reasons.UnableToGetRebalanceProfile);
+      expect(node.sendDepositTx.callCount).to.be.eq(0);
+    });
 
-  it("should request collateral without a target", async () => {
-    const assetId = AddressZero;
-    const channel = createTestChannelState("create").channel;
-    const rebalanceProfileRes = await getRebalanceProfile(channel.networkContext.chainId, assetId);
-    const profile = rebalanceProfileRes.getValue();
-
-    await requestCollateral(channel, assetId, channel.aliceIdentifier, node as INodeService, chainReader, logger);
-    node.sendDepositTx.getCall(0);
-
-    expect(
-      node.sendDepositTx.calledWith({
-        amount: BigNumber.from(profile.target).sub(getBalanceForAssetId(channel, assetId, "alice")).toString(),
-        assetId,
-        chainId: channel.networkContext.chainId,
+    it("should properly request collateral to cover payment", async () => {
+      const { channel } = createTestChannelState(UpdateType.deposit, {
+        alice: mkAddress("0xaaa"),
+        aliceIdentifier: routerPublicIdentifier,
+        assetIds: [AddressZero],
+        balances: [{ to: [mkAddress("0xaaa"), mkAddress("0xbbb")], amount: ["0", "0"] }],
+      });
+      const res = await justInTimeCollateral(
+        channel,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader,
+        log,
+        transferAmount.toString(),
+      );
+      expect(res.getError()).to.be.undefined;
+      expect(res.getValue().channelAddress).to.be.ok;
+      expect(node.sendDepositTx.callCount).to.be.eq(1);
+      expect(node.sendDepositTx.firstCall.args[0]).to.be.deep.eq({
         channelAddress: channel.channelAddress,
-        publicIdentifier: channel.aliceIdentifier,
-      }),
-    ).to.be.true;
-  });
-
-  it("should request collateral with a requested amount higher than the target", async () => {
-    const assetId = AddressZero;
-    const channel = createTestChannelState("create").channel;
-    const rebalanceProfileRes = await getRebalanceProfile(channel.networkContext.chainId, assetId);
-    const profile = rebalanceProfileRes.getValue();
-
-    const requestedAmount = BigNumber.from(profile.target).add(10);
-    await requestCollateral(
-      channel,
-      assetId,
-      channel.aliceIdentifier,
-      node as INodeService,
-      chainReader,
-      logger,
-      requestedAmount.toString(),
-    );
-    node.sendDepositTx.getCall(0);
-
-    expect(
-      node.sendDepositTx.calledWith({
-        amount: BigNumber.from(requestedAmount).sub(getBalanceForAssetId(channel, assetId, "alice")).toString(),
-        assetId,
+        publicIdentifier: routerPublicIdentifier,
+        assetId: AddressZero,
         chainId: channel.networkContext.chainId,
-        channelAddress: channel.channelAddress,
-        publicIdentifier: channel.aliceIdentifier,
-      }),
-    ).to.be.true;
+        amount: transferAmount.add(ethProfile.target).toString(),
+      });
+    });
   });
 
-  it("should request collateral with a requested amount lower than the target", async () => {
-    const assetId = AddressZero;
-    const channel = createTestChannelState("create").channel;
-    const rebalanceProfileRes = await getRebalanceProfile(channel.networkContext.chainId, assetId);
-    const profile = rebalanceProfileRes.getValue();
+  describe("adjustCollateral", () => {
+    it("should do nothing if collateralThreshold < channelBalance < reclaimThreshold", async () => {
+      const { channel } = createTestChannelState(UpdateType.deposit, {
+        alice: mkAddress("0xaaa"),
+        aliceIdentifier: routerPublicIdentifier,
+        assetIds: [AddressZero],
+        balances: [{ to: [mkAddress("0xaaa"), mkAddress("0xbbb")], amount: [ethProfile.target.toString(), "0"] }],
+      });
+      node.getStateChannel.resolves(Result.ok(channel));
+      const res = await adjustCollateral(
+        channel.channelAddress,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader,
+        log,
+      );
+      expect(res.getError()).to.be.undefined;
+      expect(res.getValue()).to.be.undefined;
+      expect(node.sendDepositTx.callCount).to.be.eq(0);
+      expect(node.withdraw.callCount).to.be.eq(0);
+    });
 
-    const requestedAmount = BigNumber.from(profile.target).sub(10);
-    await requestCollateral(
-      channel,
-      assetId,
-      channel.aliceIdentifier,
-      node as INodeService,
-      chainReader,
-      logger,
-      requestedAmount.toString(),
-    );
-    node.sendDepositTx.getCall(0);
-
-    expect(
-      node.sendDepositTx.calledWith({
-        amount: BigNumber.from(requestedAmount).sub(getBalanceForAssetId(channel, assetId, "alice")).toString(),
-        assetId,
+    it("should requestCollateral if channelBalance <= collateralizeThreshold", async () => {
+      const { channel } = createTestChannelState(UpdateType.deposit, {
+        alice: mkAddress("0xaaa"),
+        aliceIdentifier: routerPublicIdentifier,
+        assetIds: [AddressZero],
+        balances: [{ to: [mkAddress("0xaaa"), mkAddress("0xbbb")], amount: ["0", "0"] }],
+      });
+      node.getStateChannel.resolves(Result.ok(channel));
+      const res = await adjustCollateral(
+        channel.channelAddress,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader,
+        log,
+      );
+      expect(res.getError()).to.be.undefined;
+      expect(res.getValue().channelAddress).to.be.ok;
+      expect(node.sendDepositTx.callCount).to.be.eq(1);
+      expect(node.withdraw.callCount).to.be.eq(0);
+      expect(node.sendDepositTx.firstCall.args[0]).to.be.deep.eq({
+        channelAddress: channel.channelAddress,
+        publicIdentifier: routerPublicIdentifier,
+        assetId: AddressZero,
         chainId: channel.networkContext.chainId,
+        amount: ethProfile.target,
+      });
+    });
+
+    it("should reclaim if channelBalance >= reclaimThreshold", async () => {
+      const routerBalance = BigNumber.from(ethProfile.reclaimThreshold).mul(2);
+      const { channel } = createTestChannelState(UpdateType.deposit, {
+        alice: mkAddress("0xaaa"),
+        aliceIdentifier: routerPublicIdentifier,
+        assetIds: [AddressZero],
+        balances: [
+          {
+            to: [mkAddress("0xaaa"), mkAddress("0xbbb")],
+            amount: [routerBalance.toString(), "0"],
+          },
+        ],
+      });
+      node.getStateChannel.resolves(Result.ok(channel));
+      node.withdraw.resolves(Result.ok({ channelAddress: channel.channelAddress, transferId: getRandomBytes32() }));
+      const res = await adjustCollateral(
+        channel.channelAddress,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader,
+        log,
+      );
+      expect(res.getError()).to.be.undefined;
+      expect(res.getValue().channelAddress).to.be.ok;
+      expect(node.sendDepositTx.callCount).to.be.eq(0);
+      expect(node.withdraw.callCount).to.be.eq(1);
+      expect(node.withdraw.firstCall.args[0]).to.be.deep.eq({
         channelAddress: channel.channelAddress,
-        publicIdentifier: channel.aliceIdentifier,
-      }),
-    ).to.be.true;
+        publicIdentifier: routerPublicIdentifier,
+        assetId: AddressZero,
+        amount: routerBalance.sub(ethProfile.target).toString(),
+        recipient: channel.alice,
+      });
+    });
+
+    it("should fail if reclaiming fails", async () => {
+      const routerBalance = BigNumber.from(ethProfile.reclaimThreshold).mul(2);
+      const { channel } = createTestChannelState(UpdateType.deposit, {
+        alice: mkAddress("0xaaa"),
+        aliceIdentifier: routerPublicIdentifier,
+        assetIds: [AddressZero],
+        balances: [
+          {
+            to: [mkAddress("0xaaa"), mkAddress("0xbbb")],
+            amount: [routerBalance.toString(), "0"],
+          },
+        ],
+      });
+      node.getStateChannel.resolves(Result.ok(channel));
+      node.withdraw.resolves(Result.fail(new Error("fail") as any));
+      const res = await adjustCollateral(
+        channel.channelAddress,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader,
+        log,
+      );
+      expect(res.getError().message).to.be.eq(CollateralError.reasons.UnableToReclaim);
+      expect(res.getError().context).to.be.deep.eq({
+        assetId: AddressZero,
+        channelAddress: channel.channelAddress,
+        withdrawError: "fail",
+        withdrawContext: undefined,
+      });
+      expect(node.sendDepositTx.callCount).to.be.eq(0);
+      expect(node.withdraw.callCount).to.be.eq(1);
+    });
   });
 
-  it("should error on request collateral with a requested amount higher than reclaim threshold", async () => {
-    const assetId = AddressZero;
-    const channel = createTestChannelState("create").channel;
-    const rebalanceProfileRes = await getRebalanceProfile(channel.networkContext.chainId, assetId);
-    const profile = rebalanceProfileRes.getValue();
+  describe("requestCollateral", () => {
+    it("should fail if getRebalanceProfile fails", async () => {
+      getRebalanceProfile.returns(Result.fail(new Error("fail")));
+      const { channel } = createTestChannelState(UpdateType.deposit);
+      const res = await requestCollateral(
+        channel,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader,
+        log,
+      );
+      expect(res.getError().message).to.be.eq(CollateralError.reasons.UnableToGetRebalanceProfile);
+    });
 
-    const requestedAmount = BigNumber.from(profile.reclaimThreshold).add(10);
-    const req = await requestCollateral(
-      channel,
-      assetId,
-      channel.aliceIdentifier,
-      node as INodeService,
-      chainReader,
-      logger,
-      requestedAmount.toString(),
-    );
-    expect(req.isError).to.be.true;
-  });
+    it("should fail if requestedAmount > reclaimThreshold", async () => {
+      const { channel } = createTestChannelState(UpdateType.deposit);
+      const res = await requestCollateral(
+        channel,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader,
+        log,
+        BigNumber.from(ethProfile.reclaimThreshold).add(120).toString(),
+      );
+      expect(res.getError().message).to.be.eq(CollateralError.reasons.TargetHigherThanThreshold);
+    });
 
-  it("should request collateral with a transfer amount higher than the threshold", async () => {
-    const assetId = AddressZero;
-    const channel = createTestChannelState("create").channel;
-    const rebalanceProfileRes = await getRebalanceProfile(channel.networkContext.chainId, assetId);
-    const profile = rebalanceProfileRes.getValue();
+    it("should fail if it cannot get the chainProviders", async () => {
+      const { channel } = createTestChannelState(UpdateType.deposit);
+      chainReader.getHydratedProviders.returns(Result.fail(new Error("fail") as any));
+      const res = await requestCollateral(
+        channel,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader,
+        log,
+      );
+      expect(res.getError().message).to.be.eq(CollateralError.reasons.ProviderNotFound);
+    });
 
-    const requestedAmount = BigNumber.from(profile.target).sub(10);
-    const transferAmount = BigNumber.from(profile.reclaimThreshold).add(10);
-    await requestCollateral(
-      channel,
-      assetId,
-      channel.aliceIdentifier,
-      node as INodeService,
-      chainReader,
-      logger,
-      requestedAmount.toString(),
-      transferAmount.toString(),
-    );
-    node.sendDepositTx.getCall(0);
+    it("should fail if it cannot get a provider on the right chain", async () => {
+      const { channel } = createTestChannelState(UpdateType.deposit);
+      chainReader.getHydratedProviders.returns(Result.ok({ [7]: {} as any }));
+      const res = await requestCollateral(
+        channel,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader,
+        log,
+      );
+      expect(res.getError().message).to.be.eq(CollateralError.reasons.ProviderNotFound);
+    });
 
-    expect(
-      node.sendDepositTx.calledWith({
-        amount: BigNumber.from(requestedAmount)
-          .add(transferAmount)
-          .sub(getBalanceForAssetId(channel, assetId, "alice"))
-          .toString(),
-        assetId,
-        chainId: channel.networkContext.chainId,
-        channelAddress: channel.channelAddress,
-        publicIdentifier: channel.aliceIdentifier,
-      }),
-    ).to.be.true;
+    it("should fail if it cannot get the onchain balance", async () => {
+      const { channel } = createTestChannelState(UpdateType.deposit);
+      chainReader.getTotalDepositedB.resolves(Result.fail(new Error("fail") as any));
+      const res = await requestCollateral(
+        channel,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader,
+        log,
+      );
+      expect(res.getError().message).to.be.eq(CollateralError.reasons.CouldNotGetOnchainDeposits);
+    });
+
+    describe("should work", () => {
+      it("if requestedAmount is provided (and higher than target)", async () => {
+        const { channel } = createTestChannelState(UpdateType.deposit);
+        const requestedAmount = BigNumber.from(ethProfile.target).add(10000);
+        const res = await requestCollateral(
+          channel,
+          AddressZero,
+          routerPublicIdentifier,
+          node as INodeService,
+          chainReader,
+          log,
+          requestedAmount.toString(),
+        );
+        expect(res.isError).to.be.false;
+        expect(node.sendDepositTx.callCount).to.be.eq(1);
+        expect(node.sendDepositTx.firstCall.args[0]).to.be.deep.eq({
+          publicIdentifier: routerPublicIdentifier,
+          channelAddress: channel.channelAddress,
+          chainId: channel.networkContext.chainId,
+          assetId: AddressZero,
+          amount: requestedAmount.sub(channel.balances[0].amount[1]).toString(),
+        });
+        expect(node.reconcileDeposit.callCount).to.be.eq(1);
+      });
+
+      it("if requestedAmount is provided (and lower than target)", async () => {
+        const { channel } = createTestChannelState(UpdateType.deposit);
+        const requestedAmount = BigNumber.from(ethProfile.target).sub(10000);
+        const res = await requestCollateral(
+          channel,
+          AddressZero,
+          routerPublicIdentifier,
+          node as INodeService,
+          chainReader,
+          log,
+          requestedAmount.toString(),
+        );
+        expect(res.isError).to.be.false;
+        expect(node.sendDepositTx.callCount).to.be.eq(1);
+        expect(node.sendDepositTx.firstCall.args[0]).to.be.deep.eq({
+          publicIdentifier: routerPublicIdentifier,
+          channelAddress: channel.channelAddress,
+          chainId: channel.networkContext.chainId,
+          assetId: AddressZero,
+          amount: requestedAmount.sub(channel.balances[0].amount[1]).toString(),
+        });
+        expect(node.reconcileDeposit.callCount).to.be.eq(1);
+      });
+
+      it("if requestedAmount is not provided", async () => {
+        const { channel } = createTestChannelState(UpdateType.deposit);
+        const res = await requestCollateral(
+          channel,
+          AddressZero,
+          routerPublicIdentifier,
+          node as INodeService,
+          chainReader,
+          log,
+        );
+        expect(res.isError).to.be.false;
+        expect(node.sendDepositTx.callCount).to.be.eq(1);
+        expect(node.sendDepositTx.firstCall.args[0]).to.be.deep.eq({
+          publicIdentifier: routerPublicIdentifier,
+          channelAddress: channel.channelAddress,
+          chainId: channel.networkContext.chainId,
+          assetId: AddressZero,
+          amount: BigNumber.from(ethProfile.target).sub(channel.balances[0].amount[1]).toString(),
+        });
+        expect(node.reconcileDeposit.callCount).to.be.eq(1);
+      });
+
+      it("if no collateral needed", async () => {
+        const { channel } = createTestChannelState(UpdateType.deposit, {
+          balances: [
+            { to: [mkAddress(), mkAddress()], amount: [parseEther("10").toString(), parseEther("10").toString()] },
+          ],
+        });
+        const res = await requestCollateral(
+          channel,
+          AddressZero,
+          routerPublicIdentifier,
+          node as INodeService,
+          chainReader,
+          log,
+        );
+        expect(res.isError).to.be.false;
+        expect(res.getValue()).to.be.undefined;
+        expect(node.sendDepositTx.callCount).to.be.eq(0);
+        expect(node.reconcileDeposit.callCount).to.be.eq(0);
+      });
+
+      it("if there is only offchain reconciliation needed (no deposit sent onchain)", async () => {
+        const { channel } = createTestChannelState(UpdateType.deposit);
+        chainReader.getTotalDepositedB.resolves(Result.ok(parseEther("10")));
+        const res = await requestCollateral(
+          channel,
+          AddressZero,
+          routerPublicIdentifier,
+          node as INodeService,
+          chainReader,
+          log,
+        );
+        expect(res.isError).to.be.false;
+        expect(node.sendDepositTx.callCount).to.be.eq(0);
+        expect(node.reconcileDeposit.callCount).to.be.eq(1);
+      });
+    });
   });
 });