+++ modules/router/src/test/forwarding.spec.ts
@@ -34,7 +34,7 @@ import { config } from "../config";
 import * as swapService from "../services/swap";
 import * as transferService from "../services/transfer";
 import { ForwardTransferError } from "../errors";
-import * as collateralService from "../collateral";
+import * as collateralService from "../services/collateral";
 
 const testName = "Forwarding";
 
@@ -57,7 +57,7 @@ describe("Forwarding", () => {
     let receiverChannel: FullChannelState;
     let getSwappedAmount: Sinon.SinonStub;
     let cancelTransfer: Sinon.SinonStub;
-    let requestCollateral: Sinon.SinonStub;
+    let justInTimeCollateral: Sinon.SinonStub;
 
     const routerPublicIdentifier = mkPublicIdentifier("vectorRRR");
     const aliceIdentifier = mkPublicIdentifier("vectorA");
@@ -113,11 +113,13 @@ describe("Forwarding", () => {
       // get sender channel
       node.getStateChannel.onFirstCall().resolves(Result.ok(senderChannel));
       // get swapped amount (optional)
-      getSwappedAmount.resolves(Result.ok(senderTransfer.balance.amount[0]));
+      getSwappedAmount.returns(Result.ok(senderTransfer.balance.amount[0]));
       // get receiver channel
       node.getStateChannelByParticipants.onFirstCall().resolves(Result.ok(receiverChannel));
+      // check online
+      node.sendIsAliveMessage.resolves(Result.ok({ channelAddress: receiverChannel.channelAddress }));
       // request collateral (optional)
-      requestCollateral.resolves(Result.ok(undefined));
+      justInTimeCollateral.resolves(Result.ok(undefined));
       cancelTransfer.resolves(Result.ok({ channelAddress: receiverChannel.channelAddress }));
       // create receiver transfer
       node.conditionalTransfer.onFirstCall().resolves(
@@ -162,7 +164,6 @@ describe("Forwarding", () => {
     ) => {
       const { senderTransfer, receiverChannel, event } = ctx;
       expect(result.getError()).to.be.undefined;
-      console.log("result.getValue() ==========> : ", result.getValue());
       expect(result.getValue()).to.containSubset({
         channelAddress: receiverChannel.channelAddress,
         routingId: senderTransfer.meta.routingId,
@@ -256,7 +257,7 @@ describe("Forwarding", () => {
 
       chainReader = Sinon.createStubInstance(VectorChainReader);
 
-      requestCollateral = Sinon.stub(collateralService, "requestCollateral");
+      justInTimeCollateral = Sinon.stub(collateralService, "justInTimeCollateral");
     });
 
     afterEach(() => {
@@ -488,7 +489,7 @@ describe("Forwarding", () => {
       ctx.receiverChannel.networkContext.chainId = 1338;
       ctx.senderTransfer.meta.path[0].recipientChainId = 1338;
       const mocked = prepEnv(ctx);
-      getSwappedAmount.resolves(Result.fail(new Error("fail")));
+      getSwappedAmount.returns(Result.fail(new Error("fail")));
 
       const result = await forwardTransferCreation(
         mocked.event,
@@ -547,9 +548,9 @@ describe("Forwarding", () => {
       });
     });
 
-    it("fails with cancellation if transferWithAutoCollateralization indicates sender-side should be cancelled", async () => {});
+    it.skip("fails with cancellation if transferWithAutoCollateralization indicates sender-side should be cancelled", async () => {});
 
-    it("fails without cancellation if transferWithAutoCollateralization got receiver timeout", async () => {});
+    it.skip("fails without cancellation if transferWithAutoCollateralization got receiver timeout", async () => {});
 
     // TODO: the code indicates that sender should not be cancelled, verify this with Layne
     it.skip("fails with cancellation if transfer creation fails", async () => {
@@ -574,14 +575,12 @@ describe("Forwarding", () => {
     });
   });
 
-  describe("forwardTransferResolution", () => {
+  describe.skip("forwardTransferResolution", () => {
     it("should fail if it cannot find the transfers", async () => {});
     it("should fail if it cannot find incoming transfer", async () => {});
     it("should fail + queue update if resolveTransfer fails", async () => {});
     it("should work", async () => {});
   });
 
-  describe("handleIsAlive", () => {
-    it("should ");
-  });
+  describe.skip("handleIsAlive", () => {});
 });